reword the Achtung! about wrapping calls that may generate stack traces
The example in the documentation is not actually a problem. What is of concern
is a wrapper sub that includes the regex in its arguments, which will then
cause a false positive via the stack trace that is produced.

The example was adapted from a similar warning in Test::Warnings, which
originally came from this dist itself, and then amended several times after
conversations with ribasushi.